+++ github/modals/profileShareModal.ts
@@ -22,7 +22,7 @@ export async function shareProfileModal({
     uikitcontext?: UIKitInteractionContext
 }) : Promise<IUIKitModalViewParam> {
 
-    const viewId = "ProfileShareView";
+    const viewId = ModalsEnum.SHARE_PROFILE_EXEC;
     const block = modify.getCreator().getBlockBuilder();
 
     block.addActionsBlock({
@@ -90,15 +90,6 @@ export async function shareProfileModal({
                     text: 'Select Property to Share',
                   },
               }),
-              block.newButtonElement({
-                actionId : ModalsEnum.SHARE_PROFILE_EXEC,
-                text : {
-                    text : "Share to Chat",
-                    type : TextObjectType.PLAINTEXT
-                },
-                value : "shareChat"
-
-              })
         ]
     })
 
@@ -108,7 +99,15 @@ export async function shareProfileModal({
             type: TextObjectType.PLAINTEXT,
             text: "Share Profile"
         },
-        blocks: block.getBlocks()
+        blocks: block.getBlocks(),
+        submit:block.newButtonElement({
+            text : {
+                text : "Share to Chat",
+                type : TextObjectType.PLAINTEXT
+            },
+            value : "shareChat"
+
+          })
     }
 
 }
